+++ app/main.js
@@ -98,6 +98,18 @@ eval("__webpack_require__.r(__webpack_exports__);\nfunction _classCallCheck(inst
 
 /***/ }),
 
+/***/ "./src/avo/avo-adventure.js":
+/*!**********************************!*\
+  !*** ./src/avo/avo-adventure.js ***!
+  \**********************************/
+/*! exports provided: default */
+/***/ (function(module, __webpack_exports__, __webpack_require__) {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _action_pane__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./action-pane */ \"./src/avo/action-pane/index.js\");\n/* harmony import */ var _story__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./story */ \"./src/avo/story.js\");\n/* harmony import */ var _constants__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./constants */ \"./src/avo/constants.js\");\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\n\n\n\n\nvar AvoAdventure =\n/*#__PURE__*/\nfunction () {\n  function AvoAdventure(story) {\n    _classCallCheck(this, AvoAdventure);\n\n    this.mode = _constants__WEBPACK_IMPORTED_MODULE_2__[\"MODES\"].INITIALISING;\n    this.state = '';\n    this.actors = {};\n    this.assets = {};\n    this.actionPane = new _action_pane__WEBPACK_IMPORTED_MODULE_0__[\"default\"](this);\n    this.input = {\n      keysPressed: {}\n    }; // Initialise the story\n\n    this.story = story || new _story__WEBPACK_IMPORTED_MODULE_1__[\"default\"](this);\n    this.nextPlay = null;\n    this.play();\n  }\n\n  _createClass(AvoAdventure, [{\n    key: \"play\",\n    value: function play() {\n      var story = this.story;\n      story.prePlay();\n\n      if (!story.skipPlay()) {// ...\n      }\n\n      story.postPlay(); // Play and Paint are tied. \n\n      this.paint(); // Next step\n\n      this.nextPlay = setTimeout(this.play.bind(this), 1000 / _constants__WEBPACK_IMPORTED_MODULE_2__[\"FRAMES_PER_SECOND\"]);\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var story = this.story;\n\n      if (this.mode === _constants__WEBPACK_IMPORTED_MODULE_2__[\"MODES\"].ACTION) {\n        story.prePaint();\n\n        if (!story.skipPaint()) {// ...\n        }\n\n        story.postPaint();\n      }\n    }\n  }]);\n\n  return AvoAdventure;\n}();\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (AvoAdventure);\n\n//# sourceURL=webpack:///./src/avo/avo-adventure.js?");
+
+/***/ }),
+
 /***/ "./src/avo/constants.js":
 /*!******************************!*\
   !*** ./src/avo/constants.js ***!
@@ -118,7 +130,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export (binding) *
 /***/ (function(module, __webpack_exports__, __webpack_require__) {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _action_pane__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./action-pane */ \"./src/avo/action-pane/index.js\");\n/* harmony import */ var _story__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./story */ \"./src/avo/story.js\");\n/* harmony import */ var _constants__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./constants */ \"./src/avo/constants.js\");\nfunction _classCallCheck(instance, Constructor) { if (!(instance instanceof Constructor)) { throw new TypeError(\"Cannot call a class as a function\"); } }\n\nfunction _defineProperties(target, props) { for (var i = 0; i < props.length; i++) { var descriptor = props[i]; descriptor.enumerable = descriptor.enumerable || false; descriptor.configurable = true; if (\"value\" in descriptor) descriptor.writable = true; Object.defineProperty(target, descriptor.key, descriptor); } }\n\nfunction _createClass(Constructor, protoProps, staticProps) { if (protoProps) _defineProperties(Constructor.prototype, protoProps); if (staticProps) _defineProperties(Constructor, staticProps); return Constructor; }\n\n\n\n\n\nvar AvoAdventure =\n/*#__PURE__*/\nfunction () {\n  function AvoAdventure(story) {\n    _classCallCheck(this, AvoAdventure);\n\n    this.mode = _constants__WEBPACK_IMPORTED_MODULE_2__[\"MODES\"].INITIALISING;\n    this.state = '';\n    this.actors = {};\n    this.assets = {};\n    this.actionPane = new _action_pane__WEBPACK_IMPORTED_MODULE_0__[\"default\"](this);\n    this.input = {\n      keysPressed: {}\n    }; // Initialise the story\n\n    this.story = story || new _story__WEBPACK_IMPORTED_MODULE_1__[\"default\"](this);\n    this.nextPlay = null;\n    this.play();\n  }\n\n  _createClass(AvoAdventure, [{\n    key: \"play\",\n    value: function play() {\n      var story = this.story;\n      story.prePlay();\n\n      if (!story.skipPlay()) {// ...\n      }\n\n      story.postPlay(); // Play and Paint are tied. \n\n      this.paint(); // Next step\n\n      this.nextPlay = setTimeout(this.play.bind(this), 1000 / _constants__WEBPACK_IMPORTED_MODULE_2__[\"FRAMES_PER_SECOND\"]);\n    }\n  }, {\n    key: \"paint\",\n    value: function paint() {\n      var story = this.story;\n\n      if (this.mode === _constants__WEBPACK_IMPORTED_MODULE_2__[\"MODES\"].ACTION) {\n        story.prePaint();\n\n        if (!story.skipPaint()) {// ...\n        }\n\n        story.postPaint();\n      }\n    }\n  }]);\n\n  return AvoAdventure;\n}();\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (AvoAdventure);\n\n//# sourceURL=webpack:///./src/avo/index.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony import */ var _avo_adventure__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./avo-adventure */ \"./src/avo/avo-adventure.js\");\n\n/* harmony default export */ __webpack_exports__[\"default\"] = (_avo_adventure__WEBPACK_IMPORTED_MODULE_0__[\"default\"]);\n\n//# sourceURL=webpack:///./src/avo/index.js?");
 
 /***/ }),
