+++ src/avo/image-asset.js
@@ -1,0 +1,12 @@
+function ImageAsset(url) {
+  this.url = url;
+  this.img = null;
+  this.loaded = false;
+  this.img = new Image();
+  this.img.onload = function() {
+    this.loaded = true;
+  }.bind(this);
+  this.img.src = this.url;
+}
+
+export default ImageAsset;
